also remove dependency from png and xgif libs
