+++ lib/xgif/xgif.go
@@ -15,27 +15,17 @@ import (
 	"image/color"
 	"image/gif"
 	"image/png"
-	"time"
 
 	"github.com/ericpauley/go-quantize/quantize"
 
-	"oss.terrastruct.com/d2/lib/background"
 	"oss.terrastruct.com/util-go/go2"
-	"oss.terrastruct.com/util-go/xmain"
 )
 
 const INFINITE_LOOP = 0
 
 var BG_COLOR = color.White
 
-func AnimatePNGs(ms *xmain.State, pngs [][]byte, animIntervalMs int) ([]byte, error) {
-	if ms != nil {
-		cancel := background.Repeat(func() {
-			ms.Log.Info.Printf("generating GIF...")
-		}, time.Second*5)
-		defer cancel()
-	}
-
+func AnimatePNGs(pngs [][]byte, animIntervalMs int) ([]byte, error) {
 	var width, height int
 	pngImgs := make([]image.Image, len(pngs))
 	for i, pngBytes := range pngs {
